# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "This is an API to manage customers, mechanics, inventory, and service tickets in a mechanic shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /tickets/:
    post:
      tags: [Tickets]
      summary: "Create a new service ticket"
      description: "Create a new service ticket in the system"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/NewServiceTicket'
      responses:
        201:
          description: "Service ticket created"
          schema:
            $ref: '#/definitions/ServiceTicket'
        400:
          description: "Invalid input"
    get:
      tags: [Tickets]
      summary: "Get all service tickets"
      description: "Get all service tickets in the system"
      responses:
        200:
          description: "List of service tickets"
          schema:
            $ref: '#/definitions/TicketsList'
        400:
          description: "Invalid pagination parameters"
  /tickets/<int:id>:
    get:
      tags: [Tickets]
      summary: "Get a service ticket by ID"
      description: "Get service ticket in the system by ID"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Service ticket data"
          schema:
            $ref: '#/definitions/ServiceTicket'
        404:
          description: "Service ticket not found"
    put:
      tags: [Tickets]
      summary: "Update a service ticket by ID"
      description: "Update service ticket in the system by ID"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateServiceTicket'
      responses:
        200:
          description: "Service ticket updated"
        400:
          description: "Invalid input"
        404:
          description: "Service ticket not found"
  tickets/<int:id>/edit:
    put:
      tags: [Tickets]
      summary: "Update ticket mechanics"
      description: "Update a mechanic in a service ticket"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateMechanics'
      responses:
        200:
          description: "Mechanic assigned to ticket"
        400:
          description: "Invalid input"
        404:
          description: "Ticket or mechanic not found"
  /tickets/<int:id>/add_part:
    put:
      tags: [Tickets]
      summary: "Add an inventory part to a service ticket"
      description: "Update an inventory part into the service ticket"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AddPart'
      responses:
        200:
          description: "Part added to ticket"
        400:
          description: "Invalid input"
        404:
          description: "Ticket or part not found"
definitions:
...
  NewServiceTicket:
    type: object
    required:
      - VIN
      - service_date
      - service_desc
      - customer_id
      - mechanic_ids
    properties:
      VIN:
        type: string
      service_date:
        type: string
        format: date
      service_desc:
        type: string
      customer_id:
        type: integer
      mechanic_ids: 
        type: array
        items:
          type: integer
  UpdateMechanics:
    type: object
    properties:
      add_ids: 
        type: array
        items:
          type: integer
      remove_ids: 
        type: array
        items:
          type: integer
  AddPart:
    type: object
    required:
      - part_id
    properties:
      part_id:
        type: integer
  Mechanic:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number
  InventoryPart:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      price:
        type: number
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
      VIN:
        type: string
      service_date:
        type: string
        format: date
      service_desc:
        type: string
      customer_id:
        type: integer
      mechanics:
        type: array
        items:
          $ref: '#/definitions/Mechanic'
      parts:
        type: array
        items:
          $ref: '#/definitions/InventoryPart'
  # List responses
...
  TicketsList:
    type: array
    items: { $ref: '#/definitions/ServiceTicket' }